feat(#1707): extract the NodeType compile toolchain into MeshWeaver.Compiler — the framework identity pins the toolchain, not the graph
Everything that shapes a dynamic NodeType compile's GENERATED INPUT moves below
MeshWeaver.Graph into a dedicated MeshWeaver.Compiler assembly: the skeleton
generator, CodeQueryResolver (+ CodeConventions), @@-include shaping and source
aggregation/filter/join order (NodeCompileShaping), the reference set
(CompileReferences), #r/generator shaping (GeneratorPipeline), parse/compile
options + the staged, verified emit + the emit canary (EmitPipeline), failure
diagnostics (CompileDiagnostics), the framework identity (FrameworkBuildIdentity,
now also the process-lifetime FrameworkVersion accessor), the store naming
(FileSystemAssemblyStore), and the pure adoption gate.

FullMvidAssemblies flips to [MeshWeaver.Compiler, MeshWeaver.NuGet] — the NuGet
entry closes a pre-existing hole (NuGetDirectiveParser/resolver shape compile
input but were surface-hashed only). Graph body edits stop rebaking the world;
'rebuild only when we need to' (#1696) now holds in practice. The store's
filename hash is a digest of the EMITTED bytes, so the extraction is complete by
construction: any input-shaping left behind would serve stale DLLs.

The mesh-actor half (discovery, impersonation, IoPool scheduling, status
write-backs) stays in Graph and orchestrates the toolchain — one pipeline for
the portal's on-demand compile, the batch bake, and the CI bake host. Moved
public types keep their namespace and are TypeForwardedTo from Graph so modules
compiled against earlier releases keep binding. Plugin.Build's IdentityAssembly
anchor follows the toolchain (MeshWeaver.Compiler.dll); the manifest-less local
fallback is the anchor's MVID (every persistent-store host ships a surface
manifest and resolves the s-hash even locally).

First roll after this merges mints a new identity (one-time full rebake — the
key scheme changed), after which internal-only and Graph-body merges invalidate
nothing. Docs (CiContentBake, NodeTypeCompilation incl. the corrected
HasUsableBuild rules, ReleaseProcess, Modules, BuildCoordination) and all tests
adapted.

Part of #1707 (slice 1); under #1660.
