+++ src/MeshWeaver.Compiler/AssemblyStoreExtensions.cs
