+++ src/MeshWeaver.Compiler/DynamicMeshNodeAttributeGenerator.cs
