+++ src/MeshWeaver.Compiler/EmittedArtifact.cs
