+++ src/MeshWeaver.Graph/MeshWeaver.Graph.csproj
@@ -44,6 +44,11 @@
   </ItemGroup>
 
   <ItemGroup>
+    <!-- The NodeType compile toolchain (skeleton generation, source-query resolution, emit,
+         framework identity) lives in MeshWeaver.Compiler (#1707); this project keeps the
+         mesh-actor half (discovery, impersonation, status write-backs) and type-forwards the
+         moved public types for modules compiled against earlier releases. -->
+    <ProjectReference Include="..\MeshWeaver.Compiler\MeshWeaver.Compiler.csproj" />
     <ProjectReference Include="..\MeshWeaver.ContentCollections\MeshWeaver.ContentCollections.csproj" />
     <ProjectReference Include="..\MeshWeaver.Data\MeshWeaver.Data.csproj" />
     <ProjectReference Include="..\MeshWeaver.Kernel\MeshWeaver.Kernel.csproj" />
